+++ GravatarApp/AppRoot/GravatarAppApp.swift
@@ -49,9 +49,3 @@ private func resetSwiftDataStore() {
         print("❌ Error deleting SwiftData store: Could not find store file.")
     }
 }
-
-extension ProcessInfo {
-    var isSnapshotTesting: Bool {
-        environment["SNAPSHOT_TESTING"] == "1"
-    }
-}
