+++ GravatarAppTests/ModalsTests/AboutModalTests.swift
@@ -7,10 +7,6 @@ import Testing
 @Suite(.snapshots(record: .failed, diffTool: .ksdiff))
 @MainActor
 struct AboutModalTests {
-    init() async throws {
-        setenv("SNAPSHOT_TESTING", "1", 1)
-    }
-
     @Test()
     func AboutModal() async throws {
         let modalManager = ModalPresentationManager()
@@ -29,11 +25,5 @@ struct AboutModalTests {
                 .testStrategy(userInterfaceStyle: .dark, layout: .sizeThatFits),
             ]
         )
-    }
-}
-
-extension View {
-    func fullScreenFrame() -> some View {
-        self.frame(width: ViewImageConfig.iPhone13.size?.width ?? 0, height: ViewImageConfig.iPhone13.size?.height ?? 0)
     }
 }
